// examples/demo-eos-seed-app/src/generated/schema.ts
export interface Config {
    _: 'config'
    phonecalls_enabled?: true
    default_p2p_contacts?: true
    preload_featured_stickers?: true
    ignore_phone_entities?: true
    revoke_pm_inbox?: true
    blocked_mode?: true
    pfs_enabled?: true
    date: number
    expires: number
    test_mode: boolean
    this_dc: number
    dc_options: DcOption[]
    dc_txt_domain_name: string
    chat_size_max: number
    megagroup_size_max: number
    forwarded_count_max: number
    online_update_period_ms: number
    offline_blur_timeout_ms: number
    offline_idle_timeout_ms: number
    online_cloud_timeout_ms: number
    notify_cloud_delay_ms: number
    notify_default_delay_ms: number
    push_chat_period_ms: number
    push_chat_limit: number
    saved_gifs_limit: number
    edit_time_limit: number
    revoke_time_limit: number
    revoke_pm_time_limit: number
    rating_e_decay: number
    stickers_recent_limit: number
    stickers_faved_limit: number
    channels_read_media_period: number
    tmp_sessions?: number
    pinned_dialogs_count_max: number
    pinned_infolder_count_max: number
    call_receive_timeout_ms: number
    call_ring_timeout_ms: number
    call_connect_timeout_ms: number
    call_packet_timeout_ms: number
    me_url_prefix: string
    autoupdate_url_prefix?: string
    gif_search_username?: string
    venue_search_username?: string
    img_search_username?: string
    static_maps_provider?: string
    caption_length_max: number
    message_length_max: number
    webfile_dc_id: number
    suggested_lang_code?: string
    lang_pack_version?: number
    base_lang_pack_version?: number
    disabled_features: DisabledFeature[]
}
// ...
export interface DcOption {
    _: 'dcOption'
    ipv6?: true
    media_only?: true
    tcpo_only?: true
    cdn?: true
    static?: true
    id: number
    ip_address: string
    port: number
    secret?: Buffer
}
// ...
export interface DisabledFeature {
    _: 'disabledFeature'
    feature: string
    description: string
}